requirements.txt
numpy=1.20.3
tensorflow=2.5.0
Pillow=7.0.0
Flask=2.0.1
